refactor: modularize src into focused modules
Extract monolithic main.rs (~650 lines) into 4 focused modules:

- source.rs: source URL parsing and classification (+ unit tests)
- agent.rs: agent detection, symlink targets, copy/cleanup
- install.rs: canonical target, persist skills, resolve selection
- ui.rs: interactive prompts, TTY detection, colored output

main.rs is now ~130 lines of pure CLI dispatch.
lib.rs declares modules and re-exports for backward compatibility.

All 46 tests pass, zero clippy warnings.
