+++ src/lib.rs
@@ -13,145 +13,9 @@
 //! This crate is under active development. v0.1.0 is a name reservation.
 //! See the repository for progress.
 
-use thiserror::Error;
+pub mod agent;
+pub mod install;
+pub mod source;
+pub mod ui;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct GithubRepo {
-    pub owner: String,
-    pub name: String,
-    pub subfolder: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum InstallSource {
-    Github(GithubRepo),
-    LocalPath(String),
-}
-
-#[derive(Debug, Error, PartialEq, Eq)]
-pub enum SourceParseError {
-    #[error("source must be in owner/repo format")]
-    InvalidFormat,
-    #[error("owner and repo must be non-empty")]
-    EmptySegment,
-    #[error("subfolder path must be non-empty")]
-    EmptySubfolder,
-}
-
-pub fn parse_install_source(source: &str) -> Result<InstallSource, SourceParseError> {
-    if source.starts_with("./") || source.starts_with("../") || source.starts_with('/') {
-        return Ok(InstallSource::LocalPath(source.to_string()));
-    }
-
-    parse_github_source(source).map(InstallSource::Github)
-}
-
-pub fn parse_github_source(source: &str) -> Result<GithubRepo, SourceParseError> {
-    let (repo_source, subfolder) = if let Some((repo_source, subfolder)) = source.split_once(':') {
-        if subfolder.trim().is_empty() {
-            return Err(SourceParseError::EmptySubfolder);
-        }
-        (repo_source, Some(subfolder.to_string()))
-    } else {
-        (source, None)
-    };
-
-    let mut repo = parse_github_repo(repo_source)?;
-    repo.subfolder = subfolder;
-    Ok(repo)
-}
-
-pub fn parse_github_repo(source: &str) -> Result<GithubRepo, SourceParseError> {
-    let Some((owner, name)) = source.split_once('/') else {
-        return Err(SourceParseError::InvalidFormat);
-    };
-
-    if owner.trim().is_empty() || name.trim().is_empty() {
-        return Err(SourceParseError::EmptySegment);
-    }
-
-    if source.matches('/').count() != 1 {
-        return Err(SourceParseError::InvalidFormat);
-    }
-
-    Ok(GithubRepo {
-        owner: owner.to_string(),
-        name: name.to_string(),
-        subfolder: None,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_valid_owner_repo() {
-        let repo = parse_github_repo("microsoft/skills").expect("must parse");
-        assert_eq!(repo.owner, "microsoft");
-        assert_eq!(repo.name, "skills");
-        assert_eq!(repo.subfolder, None);
-    }
-
-    #[test]
-    fn reject_missing_separator() {
-        let err = parse_github_repo("microsoft-skills").expect_err("must fail");
-        assert_eq!(err, SourceParseError::InvalidFormat);
-    }
-
-    #[test]
-    fn reject_empty_segments() {
-        let err = parse_github_repo("microsoft/").expect_err("must fail");
-        assert_eq!(err, SourceParseError::EmptySegment);
-    }
-
-    #[test]
-    fn reject_multiple_slashes() {
-        let err = parse_github_repo("owner/repo/extra").expect_err("must fail");
-        assert_eq!(err, SourceParseError::InvalidFormat);
-    }
-
-    #[test]
-    fn parse_local_relative_source() {
-        let source = parse_install_source("./skills").expect("must parse");
-        assert_eq!(source, InstallSource::LocalPath("./skills".to_string()));
-    }
-
-    #[test]
-    fn parse_local_absolute_source() {
-        let source = parse_install_source("/tmp/skills").expect("must parse");
-        assert_eq!(source, InstallSource::LocalPath("/tmp/skills".to_string()));
-    }
-
-    #[test]
-    fn parse_github_source_from_install_parser() {
-        let source = parse_install_source("owner/repo").expect("must parse");
-        assert_eq!(
-            source,
-            InstallSource::Github(GithubRepo {
-                owner: "owner".to_string(),
-                name: "repo".to_string(),
-                subfolder: None,
-            })
-        );
-    }
-
-    #[test]
-    fn parse_github_source_with_subfolder() {
-        let source = parse_install_source("owner/repo:path/to/skills").expect("must parse");
-        assert_eq!(
-            source,
-            InstallSource::Github(GithubRepo {
-                owner: "owner".to_string(),
-                name: "repo".to_string(),
-                subfolder: Some("path/to/skills".to_string()),
-            })
-        );
-    }
-
-    #[test]
-    fn reject_empty_subfolder() {
-        let err = parse_install_source("owner/repo:").expect_err("must fail");
-        assert_eq!(err, SourceParseError::EmptySubfolder);
-    }
-}
+pub use source::{InstallSource, parse_install_source};
