+++ src/main.rs
@@ -1,26 +1,17 @@
 use clap::{Parser, Subcommand, error::ErrorKind};
 use std::sync::atomic::{AtomicBool, Ordering};
 
-use upskill::{InstallSource, parse_install_source};
+use upskill::agent;
+use upskill::install;
+use upskill::source::{InstallSource, parse_install_source};
+use upskill::ui;
 
-const PROJECT_CANONICAL_TARGET: &str = ".agents/skills";
-const GLOBAL_CANONICAL_TARGET: &str = ".agents/skills";
 const EXIT_SUCCESS: i32 = 0;
 const EXIT_ERROR: i32 = 1;
 const EXIT_USAGE: i32 = 2;
 const EXIT_INTERRUPTED: i32 = 130;
 
 static INTERRUPTED: AtomicBool = AtomicBool::new(false);
-
-const AGENT_SKILL_LINKS: [&str; 7] = [
-    ".claude/skills",
-    ".github/skills",
-    ".codex/skills",
-    ".cursor/skills",
-    ".kiro/skills",
-    ".windsurf/skills",
-    ".opencode/skills",
-];
 
 #[derive(Parser, Debug)]
 #[command(name = "upskill")]
@@ -137,7 +128,7 @@ fn run_add(
     copy: bool,
     global: bool,
 ) -> i32 {
-    let canonical_target = match canonical_target(global) {
+    let canonical_target = match install::canonical_target(global) {
         Ok(path) => path,
         Err(err) => {
             eprintln!("error: {}", err);
@@ -145,7 +136,7 @@ fn run_add(
         }
     };
 
-    if let Err(err) = ensure_canonical_target(&canonical_target) {
+    if let Err(err) = install::ensure_canonical_target(&canonical_target) {
         eprintln!("error: {}", err);
         return 1;
     }
@@ -157,7 +148,7 @@ fn run_add(
             } else {
                 format!("github:{}/{}", repo.owner, repo.name)
             };
-            let resolved_skills = match resolve_requested_skills(skills, &repo.name) {
+            let resolved_skills = match install::resolve_requested_skills(skills, &repo.name) {
                 Ok(skills) => skills,
                 Err(err) => {
                     eprintln!("error: {}", err);
@@ -165,16 +156,18 @@ fn run_add(
                 }
             };
 
-            if let Err(err) =
-                persist_installed_skills(&canonical_target, &resolved_skills, &source_label)
-            {
+            if let Err(err) = install::persist_installed_skills(
+                &canonical_target,
+                &resolved_skills,
+                &source_label,
+            ) {
                 eprintln!("error: {}", err);
                 return 1;
             }
 
             if !global
                 && let Err(err) =
-                    ensure_agent_targets(claude, copilot, all, copy, &canonical_target)
+                    agent::ensure_agent_targets(claude, copilot, all, copy, &canonical_target)
             {
                 eprintln!("error: {}", err);
                 return 1;
@@ -186,7 +179,7 @@ fn run_add(
             if let Some(subfolder) = repo.subfolder {
                 println!("subfolder: {}", subfolder);
             }
-            print_selected_skills(&resolved_skills, skills.is_empty());
+            ui::print_selected_skills(&resolved_skills, skills.is_empty());
             0
         }
         Ok(InstallSource::LocalPath(path)) => {
@@ -200,7 +193,7 @@ fn run_add(
                 .and_then(|v| v.to_str())
                 .filter(|v| !v.is_empty())
                 .unwrap_or("local-skill");
-            let resolved_skills = match resolve_requested_skills(skills, default_skill) {
+            let resolved_skills = match install::resolve_requested_skills(skills, default_skill) {
                 Ok(skills) => skills,
                 Err(err) => {
                     eprintln!("error: {}", err);
@@ -209,16 +202,18 @@ fn run_add(
             };
 
             let source_label = format!("local:{}", path);
-            if let Err(err) =
-                persist_installed_skills(&canonical_target, &resolved_skills, &source_label)
-            {
+            if let Err(err) = install::persist_installed_skills(
+                &canonical_target,
+                &resolved_skills,
+                &source_label,
+            ) {
                 eprintln!("error: {}", err);
                 return 1;
             }
 
             if !global
                 && let Err(err) =
-                    ensure_agent_targets(claude, copilot, all, copy, &canonical_target)
+                    agent::ensure_agent_targets(claude, copilot, all, copy, &canonical_target)
             {
                 eprintln!("error: {}", err);
                 return 1;
@@ -226,7 +221,7 @@ fn run_add(
 
             println!("install source: local");
             println!("path: {}", path);
-            print_selected_skills(&resolved_skills, skills.is_empty());
+            ui::print_selected_skills(&resolved_skills, skills.is_empty());
             0
         }
         Err(err) => {
@@ -237,7 +232,7 @@ fn run_add(
 }
 
 fn run_list(global: bool) -> i32 {
-    let canonical = match canonical_target(global) {
+    let canonical = match install::canonical_target(global) {
         Ok(path) => path,
         Err(err) => {
             eprintln!("error: {}", err);
@@ -282,7 +277,7 @@ fn run_list(global: bool) -> i32 {
         return 0;
     }
 
-    let active_agents = detect_active_agents();
+    let active_agents = agent::detect_active_agents();
     let symlink_text = if active_agents.is_empty() {
         "none".to_string()
     } else {
@@ -297,7 +292,7 @@ fn run_list(global: bool) -> i32 {
 }
 
 fn run_remove(skill: &str, yes: bool, global: bool) -> i32 {
-    let canonical = match canonical_target(global) {
+    let canonical = match install::canonical_target(global) {
         Ok(path) => path,
         Err(err) => {
             eprintln!("error: {}", err);
@@ -311,7 +306,7 @@ fn run_remove(skill: &str, yes: bool, global: bool) -> i32 {
         return 2;
     }
 
-    if should_prompt_for_confirmation(yes) && !confirm_removal(skill) {
+    if ui::should_prompt_for_confirmation(yes) && !ui::confirm_removal(skill) {
         eprintln!("error: removal cancelled");
         return 1;
     }
@@ -321,309 +316,11 @@ fn run_remove(skill: &str, yes: bool, global: bool) -> i32 {
         return 1;
     }
 
-    if !global && let Err(err) = cleanup_agent_symlinks_if_empty(&canonical) {
+    if !global && let Err(err) = agent::cleanup_agent_symlinks_if_empty(&canonical) {
         eprintln!("error: {}", err);
         return 1;
     }
 
     println!("removed skill: {}", skill);
     0
-}
-
-fn should_prompt_for_confirmation(yes: bool) -> bool {
-    use std::io::IsTerminal;
-
-    !yes && std::io::stdin().is_terminal()
-}
-
-fn confirm_removal(skill: &str) -> bool {
-    use std::io::{self, IsTerminal, Write};
-
-    if io::stdout().is_terminal() && std::env::var_os("NO_COLOR").is_none() {
-        print!("\u{1b}[33mremove skill '{}' ? [y/N]:\u{1b}[0m ", skill);
-    } else {
-        print!("remove skill '{}' ? [y/N]: ", skill);
-    }
-    if io::stdout().flush().is_err() {
-        return false;
-    }
-
-    let mut answer = String::new();
-    if io::stdin().read_line(&mut answer).is_err() {
-        return false;
-    }
-
-    matches!(answer.trim().to_ascii_lowercase().as_str(), "y" | "yes")
-}
-
-fn cleanup_agent_symlinks_if_empty(canonical: &std::path::Path) -> Result<(), String> {
-    if !canonical_has_skills(canonical)? {
-        for link in AGENT_SKILL_LINKS {
-            remove_symlink_if_exists(link)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn canonical_has_skills(canonical: &std::path::Path) -> Result<bool, String> {
-    if !canonical.exists() {
-        return Ok(false);
-    }
-
-    let entries = std::fs::read_dir(canonical)
-        .map_err(|err| format!("failed to inspect installed skills: {}", err))?;
-
-    for entry in entries {
-        let entry = entry.map_err(|err| format!("failed to inspect installed skills: {}", err))?;
-        if entry.path().is_dir() {
-            return Ok(true);
-        }
-    }
-
-    Ok(false)
-}
-
-fn remove_symlink_if_exists(link_path: &str) -> Result<(), String> {
-    let link = std::path::Path::new(link_path);
-    match std::fs::symlink_metadata(link) {
-        Ok(meta) => {
-            if meta.file_type().is_symlink() {
-                std::fs::remove_file(link).map_err(|err| {
-                    format!("failed to remove symlink {}: {}", link.display(), err)
-                })?;
-            }
-            Ok(())
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => Ok(()),
-        Err(err) => Err(format!("failed to inspect {}: {}", link.display(), err)),
-    }
-}
-
-fn resolve_requested_skills(skills: &[String], default_skill: &str) -> Result<Vec<String>, String> {
-    if !skills.is_empty() {
-        return Ok(skills.to_vec());
-    }
-
-    if !interactive_skill_selection_enabled() {
-        return Ok(vec![default_skill.to_string()]);
-    }
-
-    prompt_for_skill_selection(default_skill)
-}
-
-fn interactive_skill_selection_enabled() -> bool {
-    use std::io::IsTerminal;
-
-    std::env::var_os("UPSKILL_FORCE_INTERACTIVE").is_some() || std::io::stdin().is_terminal()
-}
-
-fn prompt_for_skill_selection(default_skill: &str) -> Result<Vec<String>, String> {
-    use std::io::{self, Write};
-
-    print!(
-        "select skills (comma-separated, empty for '{}'): ",
-        default_skill
-    );
-    io::stdout()
-        .flush()
-        .map_err(|err| format!("failed to flush prompt: {}", err))?;
-
-    let mut answer = String::new();
-    io::stdin()
-        .read_line(&mut answer)
-        .map_err(|err| format!("failed to read selected skills: {}", err))?;
-
-    let parsed: Vec<String> = answer
-        .split(',')
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToString::to_string)
-        .collect();
-
-    if parsed.is_empty() {
-        return Ok(vec![default_skill.to_string()]);
-    }
-
-    Ok(parsed)
-}
-
-fn persist_installed_skills(
-    canonical_target: &std::path::Path,
-    skills: &[String],
-    source: &str,
-) -> Result<(), String> {
-    for skill in skills {
-        let skill_dir = canonical_target.join(skill);
-        std::fs::create_dir_all(&skill_dir)
-            .map_err(|err| format!("failed to create {}: {}", skill_dir.display(), err))?;
-        std::fs::write(skill_dir.join(".upskill-source"), source)
-            .map_err(|err| format!("failed to write source metadata for {}: {}", skill, err))?;
-    }
-
-    Ok(())
-}
-
-fn detect_active_agents() -> Vec<String> {
-    let mut agents = Vec::new();
-
-    if std::fs::symlink_metadata(".claude/skills").is_ok() {
-        agents.push("claude".to_string());
-    }
-    if std::fs::symlink_metadata(".github/skills").is_ok() {
-        agents.push("copilot".to_string());
-    }
-    if std::fs::symlink_metadata(".codex/skills").is_ok() {
-        agents.push("codex".to_string());
-    }
-    if std::fs::symlink_metadata(".cursor/skills").is_ok() {
-        agents.push("cursor".to_string());
-    }
-    if std::fs::symlink_metadata(".kiro/skills").is_ok() {
-        agents.push("kiro".to_string());
-    }
-    if std::fs::symlink_metadata(".windsurf/skills").is_ok() {
-        agents.push("windsurf".to_string());
-    }
-    if std::fs::symlink_metadata(".opencode/skills").is_ok() {
-        agents.push("opencode".to_string());
-    }
-
-    agents
-}
-
-fn print_selected_skills(skills: &[String], implicit_selection: bool) {
-    if skills.is_empty() || implicit_selection {
-        return;
-    }
-
-    println!("skills: {}", skills.join(","));
-}
-
-fn canonical_target(global: bool) -> Result<std::path::PathBuf, String> {
-    if global {
-        let home = std::env::var_os("HOME")
-            .map(std::path::PathBuf::from)
-            .ok_or_else(|| "HOME is not set".to_string())?;
-        return Ok(home.join(GLOBAL_CANONICAL_TARGET));
-    }
-
-    Ok(std::path::PathBuf::from(PROJECT_CANONICAL_TARGET))
-}
-
-fn ensure_canonical_target(canonical_target: &std::path::Path) -> Result<(), String> {
-    std::fs::create_dir_all(canonical_target).map_err(|err| {
-        format!(
-            "failed to create canonical target {}: {}",
-            canonical_target.display(),
-            err
-        )
-    })
-}
-
-fn ensure_agent_targets(
-    claude: bool,
-    copilot: bool,
-    all: bool,
-    copy: bool,
-    canonical_target: &std::path::Path,
-) -> Result<(), String> {
-    if all {
-        for link in AGENT_SKILL_LINKS {
-            create_agent_target(link, copy, canonical_target)?;
-        }
-        return Ok(());
-    }
-
-    let auto_detect = !claude && !copilot;
-
-    let link_claude = claude || (auto_detect && std::path::Path::new(".claude").exists());
-    let link_copilot = copilot || (auto_detect && std::path::Path::new(".github").exists());
-
-    if link_claude {
-        create_agent_target(".claude/skills", copy, canonical_target)?;
-    }
-
-    if link_copilot {
-        create_agent_target(".github/skills", copy, canonical_target)?;
-    }
-
-    Ok(())
-}
-
-fn create_agent_target(
-    link_path: &str,
-    copy: bool,
-    canonical_target: &std::path::Path,
-) -> Result<(), String> {
-    let link = std::path::Path::new(link_path);
-
-    if let Some(parent) = link.parent() {
-        std::fs::create_dir_all(parent)
-            .map_err(|err| format!("failed to create {}: {}", parent.display(), err))?;
-    }
-
-    if link.exists() || std::fs::symlink_metadata(link).is_ok() {
-        if link.is_dir() && !link.is_symlink() {
-            std::fs::remove_dir_all(link)
-                .map_err(|err| format!("failed to reset {}: {}", link.display(), err))?;
-        } else {
-            std::fs::remove_file(link)
-                .map_err(|err| format!("failed to reset {}: {}", link.display(), err))?;
-        }
-    }
-
-    if copy {
-        copy_dir_all(canonical_target, link)?;
-        return Ok(());
-    }
-
-    #[cfg(unix)]
-    {
-        std::os::unix::fs::symlink(canonical_target, link).map_err(|err| {
-            format!(
-                "failed to create symlink {} -> {}: {}",
-                link.display(),
-                canonical_target.display(),
-                err
-            )
-        })?;
-    }
-
-    #[cfg(not(unix))]
-    {
-        let _ = link;
-        return Err("symlink creation is currently supported on unix platforms".to_string());
-    }
-
-    Ok(())
-}
-
-fn copy_dir_all(src: &std::path::Path, dst: &std::path::Path) -> Result<(), String> {
-    std::fs::create_dir_all(dst)
-        .map_err(|err| format!("failed to create {}: {}", dst.display(), err))?;
-
-    let entries = std::fs::read_dir(src)
-        .map_err(|err| format!("failed to read {}: {}", src.display(), err))?;
-
-    for entry in entries {
-        let entry = entry.map_err(|err| format!("failed to read {}: {}", src.display(), err))?;
-        let entry_path = entry.path();
-        let target_path = dst.join(entry.file_name());
-
-        if entry_path.is_dir() {
-            copy_dir_all(&entry_path, &target_path)?;
-        } else {
-            std::fs::copy(&entry_path, &target_path).map_err(|err| {
-                format!(
-                    "failed to copy {} to {}: {}",
-                    entry_path.display(),
-                    target_path.display(),
-                    err
-                )
-            })?;
-        }
-    }
-
-    Ok(())
 }
